Extract init-args input models into single module
