+++ tests/v2/04_runner/fractal_tasks_core_mock/src/fractal_tasks_mock/input_models.py
@@ -1,0 +1,19 @@
+from typing import Literal
+from typing import Optional
+
+from pydantic import BaseModel
+from pydantic import Field
+
+
+class InitArgsRegistration(BaseModel):
+    ref_path: str
+
+
+class InitArgsCellVoyager(BaseModel):
+    raw_path: str
+    acquisition: Optional[int]
+
+
+class InitArgsIllumination(BaseModel):
+    raw_path: str
+    subsets: dict[Literal["C_index"], int] = Field(default_factory=dict)
